moving property detail api into PropertiesController.ts
