+++ src/controllers/PropertiesController.ts
@@ -1,14 +1,17 @@
 // eslint-disable spaced-comment
-import { Get, Route, Controller, TsoaResponse, Res, Path } from 'tsoa';
+import { Get, Route, Controller, TsoaResponse, Res, Path, Query } from 'tsoa';
 import logger from '../middleware/logger';
 import GeocodeAPICaller from '../helpers/geocodeAPICaller';
 import {
     geocoderReferenceErrorType,
     getAddressResults,
     pinRangeErrorType,
+    propertyDetailsResponse,
     searchRangeErrorType,
     serverErrorType,
 } from '../helpers/types';
+import { findPropertyDetails } from '../db/ActivePIN.db';
+import axios from 'axios';
 
 @Route('properties')
 export class PropertiesController extends Controller {
@@ -27,6 +30,28 @@ export class PropertiesController extends Controller {
      *  Each result contains the match score (from 50-100, 100 being a perfect match) of the result,
      *  the full address of the property, and the siteID
      */
+
+    /**
+     * Used to get property owner details from a given a site ID
+     * - Step 1: Call bc geocoder parcel API to get PIDs from site ID
+     * - Step 2: Check database and return property details for properties with a matching PID
+     * Expected error codes and messages:
+     * - 200
+     * -- 'OK'
+     * - 204
+     *  -- 'No Content
+     * - 400
+     *  -- 'Bad Request'
+     * - 401
+     * -- 'Unauthorized'
+     * - 403
+     * -- 'Forbidden'
+     * - 404
+     * -- 'Not Found'
+     * @param siteID The siteID of a site
+     * @returns An object containing the property owner details
+     */
+
     @Get('address/{address}')
     public async getSiteID(
         @Res() rangeErrorResponse: TsoaResponse<422, searchRangeErrorType>,
@@ -60,5 +85,33 @@ export class PropertiesController extends Controller {
             }
         }
         return res;
+    }
+
+    @Get('details')
+    public async getPropertyDetails(
+        @Query() siteID: string,
+    ): Promise<propertyDetailsResponse> {
+        const parcelsApiUrl = `${process.env.BCGEOCODER_TEST_API_URL_PID}`;
+        const jsonFormat = '.json';
+
+        const pid = async () => {
+            try {
+                return await axios.get(
+                    `${parcelsApiUrl}${siteID}${jsonFormat}`,
+                    {
+                        headers: {
+                            apikey: `${process.env.BCGEOCODER_API_KEY_PID}`,
+                        },
+                    },
+                );
+            } catch (err) {
+                logger.debug(err);
+                return err;
+            }
+        };
+
+        const pidsData: any = await pid();
+        const result = await findPropertyDetails(parseInt(pidsData.data.pids));
+        return result;
     }
 }
