+++ src/routes/propertyDetails.ts
@@ -1,12 +1,12 @@
 import express, { Request, Response } from 'express';
-import { PropertyDetailsController } from '../controllers/propertyDetailsController';
+import { PropertiesController } from '../controllers/PropertiesController';
 
 const propertyDetailsRouter = express.Router();
 
 propertyDetailsRouter.get(
     '/propertyDetails',
     async (req: Request, res: Response) => {
-        const controller = new PropertyDetailsController();
+        const controller = new PropertiesController();
         const response = await controller.getPropertyDetails(req.params.siteID);
         return res.send(response);
     },
